refactor: MessageManager 타입으로 메시지 관리 객체 분리
